requirements.txt
# Core web framework
flask>=3.0.0
flask-cors>=4.0.0

# Face recognition and computer vision
deepface>=0.0.91
opencv-python>=4.8.0
pillow>=10.0.0

# Video processing and downloading
yt-dlp>=2023.12.0
moviepy>=1.0.3

# Machine learning and clustering
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0

# Database and data handling
sqlalchemy>=2.0.0

# Background task processing
celery>=5.3.0
redis>=5.0.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0

# Development
pytest>=7.4.0
black>=23.11.0